+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ## 0.14.2
 Bug:
   * Fixed nil pointer issue in modify cookie
+  * Added a garbage collection on reload due to golang bug, also to track memory usage
 
 ## 0.14.1
 Bug: